dom/base/test/test_window_proto.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test for ...</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws(new TypeError, function() {
    Object.setPrototypeOf(window, Object.create(window));
  }, "Setting prototype via setPrototypeOf");

  assert_throws(new TypeError, function() {
    window.__proto__ = Object.create(window);
  }, "Setting prototype via __proto__");
}, "Setting the prototype of a window to something that has the window on its proto chain should throw");
</script>
